requirements.txt
langchain>=1.0.0
langchain-google-genai>=4.2.2
langgraph>=1.0.0
yfinance>=0.2.66
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
plotly>=5.24.0
requests>=2.32.0
python-dotenv>=1.0.1
